move the api code layout around in preparation for adding delete
