+++ app/api/jobs/route.ts
@@ -1,5 +1,14 @@
 import { createClient } from '@/lib/supabase/server';
 
+export async function GET() {
+  const supabase = await createClient();
+  const { data: jobs, error } = await supabase.from('jobs').select('*').order('id');
+  if (error) {
+    console.log(`error getting all jobs ${error}`);
+  }
+  return Response.json(jobs);
+}
+
 export async function POST(request: Request) {
   const supabase = await createClient();
